requirements.txt
aiogram>=3.4.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
loguru>=0.7.0